linebot_forpine/requirements.txt
beautifulsoup4==4.5.1
chainer==1.17.0
click==6.6
cssutils==1.0.1
Django==1.9.2
falcon==1.0.0
filelock==2.0.6
Flask==0.11.1
future==0.16.0
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
line-bot-sdk==1.0.2
lxml==3.6.4
MarkupSafe==0.23
nose==1.3.7
numpy==1.11.1
pi==0.1.2
Pillow==3.4.1
protobuf==3.1.0.post1
pycurl==7.43.0
python-firebase==1.2
python-mimeparse==1.6.0
requests==2.11.1
scipy==0.18.1
selenium==3.0.1
six==1.10.0
urllib3==1.18
Werkzeug==0.11.11
